requirements.txt
# MCP Framework
mcp[cli]>=1.0.0
fastmcp>=0.1.0

# Hyperliquid SDK
hyperliquid-python-sdk>=0.3.0

# HTTP & WebSocket
httpx>=0.24.0
websockets>=12.0

# Ethereum account utilities for signing
eth-account>=0.8.0

# Configuration & Data Validation
python-dotenv>=1.0.0
pydantic>=2.0.0

# Type Hints
typing-extensions>=4.0.0

# Development Dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0
